perf: implement dynamic tile dispatch and algorithmic I/O optimizations
Binary Format & Template Dispatching:
- Deprecated the static 20-byte STileEntry in favor of dynamic STileEntry32 (8 bytes) and STileEntry64 (16 bytes) structures.
- Optimized SMapHeader to maintain a strict 40-byte footprint by shrinking version and EMapType to uint8, introducing ETileEntryFormat, and adding a 1-byte reserved alignment pad.
- Implemented template dispatching (WriteTiles<T>) to evaluate the actual baked payload size during PrePass, falling back to 64-bit LFS structures only when the 4GB threshold is physically exceeded.

Algorithmic Optimizations:
- Replaced byte-by-byte channel extraction with block-level std::memcpy row copying in ExtractTileDataHelper.
- Flattened the nested coordinate loops (ty, tx) in WriteTiles into a single linear pass with smart coordinate increments.

+++ Core/Baking/BakeManager.h
@@ -4,7 +4,7 @@
 
 namespace JDKLevelMaps
 {
-	enum class EMapType : uint16;
+	enum class EMapType : uint8;
 }
 
 namespace JDKLevelMaps::Settings
